Fix STORYBOOK_IS_RUNNING flag, review this
