+++ app/assets/javascripts/reader_profile_january/FAndPEnglishView.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import {F_AND_P_ENGLISH} from '../reading/thresholds';
-import {interpretFAndPEnglish} from '../reading/fAndPInterpreter';
 import {COMPREHENSION} from './instructionalStrategies';
 import {matchStrategies} from './instructionalStrategies';
 import {mostRecentDataPoint} from './dibelsParsing';
 import expandedViewPropTypes from './expandedViewPropTypes';
 import ExpandedLayout from './ExpandedLayout';
-import MaterialsCarousel from './MaterialsCarousel';
 import Strategies from './Strategies';
 import Data from './Data';
+import FAndPMaterials from './FAndPMaterials';
 
 
 const benchmarkAssessmentKey = F_AND_P_ENGLISH;
@@ -41,17 +40,7 @@ export default class FAndPEnglishView extends React.Component {
     const {readerJson} = this.props;
     const dataPoint = mostRecentDataPoint(readerJson, benchmarkAssessmentKey);
     if (!dataPoint) return null;
-    const level = interpretFAndPEnglish(dataPoint.json.value);
-    if (!level) return null;
-    const fileKeys = MATERIAL_URLS[level];
-    if (!fileKeys) return null;
-
-    return (
-      <div>
-        <MaterialsCarousel fileKeys={fileKeys} />
-        <div style={styles.level}>Instructional level: {level}</div>
-      </div>
-    );
+    return <FAndPMaterials rawLevelText={dataPoint.json.value} />;
   }
 
   renderData() {
@@ -67,17 +56,4 @@ export default class FAndPEnglishView extends React.Component {
   }
 }
 FAndPEnglishView.propTypes = expandedViewPropTypes;
-
-const styles = {
-  level: {
-    paddingLeft: 5,
-    fontSize: 12
-  }
-};
-
-const MATERIAL_URLS = {
-  'A': ['FP-A1-cover', 'FP-A1-page', 'FP-A2-cover', 'FP-A2-page'],
-  'B': ['FP-B1-cover', 'FP-B1-page', 'FP-B2-cover', 'FP-B2-page']
-};
-
