+++ package.json
@@ -13,7 +13,7 @@
     "start": "concurrently --kill-others \"bundle exec rails s\" \"yarn dev\"",
     "test": "yarn jest --watch",
     "dev": "NODE_ENV=development webpack --config ui/config/webpack.dev.js --watch",
-    "storybook": "STORYBOOK_RUNNING=true start-storybook -p 6006 --config-dir ui/config/.storybook --quiet --ci",
+    "storybook": "start-storybook -p 6006 --config-dir ui/config/.storybook --quiet --ci",
     "build": "NODE_ENV=production webpack --config ui/config/webpack.prod.js",
     "test-cli": "yarn yarn-checks-cli && yarn lint-cli && yarn jest && yarn audits && yarn storybook-cli",
     "jest": "TZ='America/New_York' jest --config jest.json",
